pyproject.toml
[tool.poetry]
name = "readmewriterassistant"
version = "0.1.0"
description = ""
authors = ["Bruno Guedes <bruno@bclgapps.com.au>"]
readme = "README.md"
packages = [{include = "readmewriterassistant", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.33.0"
python-dotenv = "^1.0.1"
langchain = "^0.1.16"
langchain-community = "^0.0.34"
langchain-anthropic = "^0.1.11"
langchain-openai = "^0.1.3"
langchain-groq = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
